pyproject.toml
[tool.poetry]
name = "irmas_torch"
version = "0.1.0"
description = ""
authors = ["Cézar Lemos <cezarcbl@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.*"
torch = "^1.12.1"
pytorch-lightning = "^1.7.1"
numpy = "^1.23.2"
torchaudio = "^0.12.1"
scipy = "^1.9.0"
scikit-learn = "^1.1.2"
hydra-core = "^1.2.0"
torchvision = "^0.13.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
